model is now a class hidden in the view file
