--- spynnaker/pyNN/data/spynakker_data_model.py
@@ -1,52 +1,0 @@
-# Copyright (c) 2017-2019 The University of Manchester
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-# Does not extend FecDataModel to keep singleton management clean
-class SpynnakerDataModel(object):
-    """
-    Singleton data model
-
-    This class should not be accessed directly please use the DataView and
-    DataWriter classes.
-    Accessing or editing the data held here directly is NOT SUPPORTED
-
-    There are other DataModel classes which sit next to this one and hold
-    additional data. The DataView and DataWriter classes will combine these
-    as needed.
-
-    What data is held where and how can change without notice.
-    """
-
-    __singleton = None
-
-    __slots__ = [
-        # Data values cached
-        "__min_delay"
-    ]
-
-    def __new__(cls):
-        if cls.__singleton:
-            return cls.__singleton
-        # pylint: disable=protected-access
-        obj = object.__new__(cls)
-        cls.__singleton = obj
-        obj.__clear()
-        return obj
-
-    def __clear(self):
-        """
-        Clears out all data
-        """
-        self.__min_delay = None
